Fix guided mode category merge and delete
In the case of merging categories after training the first time make sure that
predictions are not changed. When deleting a category with predictions attached
to it warn the user that this will automatically start a retraining. Delete and
retrain after user approval. New categories with no predictions attached can be
deleted without issue.