Improve clarity for excluded labels
The "block" icon is always displayed if a label has been excluded, otherwise it
is only visible on hover. If a label is excluded the row is highlighted in red
and the font weight is changed to bold to make it clearer. The tooltip also
updates depending on the state now.