Only set up the "keyup" event listener for modes that use it
Since the component is now cached using keep-alive, the "mounted" lifecycle
hook is only called twice: once for the initial load of Guided or Labeling mode
(whichever happens first), and once in Review mode when a modified version of
the component is used. Since the Review mode does not use hotkeys we do not add
the event listener.