Do not update the review superpixels unless we're in review mode
Now that the review mode is cached after it has been viewed the first time,
when a superpixel label is changed in the guided mode this triggers a
re-compute of the filtered, sorted and grouped list used to display the
superpixels for the review mode. This can be a slow process though. For now,
simply prevent the update unless we are in the review mode.