Prevent console error from styling request
Sets a default return value in case no fill color is found. This prevents a
console error from a race condition when the grouping selection is changed but
internally the actual grouping has not changed - this request a category fill
color for "data" which is not actually one of the user created labels.